python1901/6_django/axfday13/代码/axf下午代码/carts/urls.py

from rest_framework.routers import SimpleRouter

from carts.views import CartView

router = SimpleRouter()
router.register('cart', CartView)

urlpatterns = [

]

urlpatterns += router.urls
