python1901/6_django/axfday13/代码/axf下午代码/requirement.txt
django==2.0.7
pymysql==0.9.3
djangorestframework==3.9.3
django-filter==2.1.0
django-redis==4.10.0
redis==2.10.6
django-cors-headers==2.5.3